feat: refatoracao de projeto

+++ estande_repository.py
@@ -6,7 +6,7 @@ class EstandeRepository:
         with self.conn.cursor() as cursor:
             sql = """
                 INSERT INTO estande (nome, tema, imagem, descricao, evento_id, admin_user_id)
-                VALUES (%s, %s, %s, %s, %s)
+                VALUES (%s, %s, %s, %s, %s, %s)
             """
             cursor.execute(sql, (
                 estande.nome,
@@ -18,16 +18,18 @@ class EstandeRepository:
             ))
         self.conn.commit()
 
-    def listar_todos(self):
+    def listar_todos(self, admin_user_id):
         with self.conn.cursor() as cursor:
-            cursor.execute("SELECT * FROM estande")
+            sql = "SELECT * FROM estande WHERE admin_user_id = %s"
+            cursor.execute(sql, (admin_user_id,))
             return cursor.fetchall()
 
     def atualizar(self, estande):
         with self.conn.cursor() as cursor:
             sql = """
-                UPDATE estande SET nome = %s, tema = %s, imagem = %s, descricao = %s, evento_id = %s, admin_user_id = %s
-                WHERE id = %s
+                UPDATE estande 
+                SET nome = %s, tema = %s, imagem = %s, descricao = %s, evento_id = %s
+                WHERE id = %s AND admin_user_id = %s
             """
             cursor.execute(sql, (
                 estande.nome,
@@ -35,13 +37,13 @@ class EstandeRepository:
                 estande.imagem,
                 estande.descricao,
                 estande.evento_id,
-                estande.admin_user_id,
-                estande.id
+                estande.id,
+                estande.admin_user_id
             ))
         self.conn.commit()
 
-    def deletar(self, estande_id):
+    def deletar(self, estande_id, admin_user_id):
         with self.conn.cursor() as cursor:
-            sql = "DELETE FROM estande WHERE id = %s"
-            cursor.execute(sql, (estande_id,))
+            sql = "DELETE FROM estande WHERE id = %s AND admin_user_id = %s"
+            cursor.execute(sql, (estande_id, admin_user_id))
         self.conn.commit()
