+++ evento_repository.py
@@ -6,7 +6,7 @@ class EventoRepository:
         with self.conn.cursor() as cursor:
             sql = """
                 INSERT INTO evento (nome, data_inicial, data_final, imagem, descricao, cep, rua, bairro, cidade, numero, complemento, admin_user_id)
-                VALUES (%s, %s, %s, %s, %s, %s)
+                VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)
             """
             cursor.execute(sql, (
                 evento.nome,
@@ -24,16 +24,18 @@ class EventoRepository:
             ))
         self.conn.commit()
 
-    def listar_todos(self):
+    def listar_todos(self, admin_user_id):
         with self.conn.cursor() as cursor:
-            cursor.execute("SELECT * FROM evento")
+            sql = "SELECT * FROM evento WHERE admin_user_id = %s"
+            cursor.execute(sql, (admin_user_id,))
             return cursor.fetchall()
 
     def atualizar(self, evento):
         with self.conn.cursor() as cursor:
             sql = """
-                UPDATE evento SET nome = %s, data_inicial = %s, data_final = %s, imagem = %s, descricao = %s, cep = %s, rua = %s, bairro = %s, cidade = %s, numero = %s, complemento = %s, admin_user_id = %s
-                WHERE id = %s
+                UPDATE evento 
+                SET nome = %s, data_inicial = %s, data_final = %s, imagem = %s, descricao = %s, cep = %s, rua = %s, bairro = %s, cidade = %s, numero = %s, complemento = %s
+                WHERE id = %s AND admin_user_id = %s
             """
             cursor.execute(sql, (
                 evento.nome,
@@ -47,13 +49,13 @@ class EventoRepository:
                 evento.cidade,
                 evento.numero,
                 evento.complemento,
-                evento.admin_user_id,
-                evento.id
+                evento.id,
+                evento.admin_user_id
             ))
         self.conn.commit()
 
-    def deletar(self, evento_id):
+    def deletar(self, evento_id, admin_user_id):
         with self.conn.cursor() as cursor:
-            sql = "DELETE FROM evento WHERE id = %s"
-            cursor.execute(sql, (evento_id,))
+            sql = "DELETE FROM evento WHERE id = %s AND admin_user_id = %s"
+            cursor.execute(sql, (evento_id, admin_user_id))
         self.conn.commit()
